+++ tests/TestJSON.cpp
@@ -6,7 +6,7 @@
 #include <string>
 #include <memory>
 
-#include "Database/json.h"
+#include "Tools/json.h"
 
 int main ( int argc, char * argv [] ) {
   using namespace JSON;
